feat: extract yield endpoints into dedicated YieldService and package, wire into NotbankClient
Move yield-related functionality out of Wallet/Product services into a new exchange.notbank.yield package:

- Add YieldService with methods for depositToYield, withdrawFromYield and getYieldProducts.
- Introduce YieldServiceResponseAdapter and yield-specific Endpoints, param builders and response models.
- Wire YieldService into NotbankClient and NotbankClientFactory so clients can obtain it.
- Remove yield-specific endpoints/adapters from WalletService, WalletServiceResponseAdapter and ProductService; tests updated accordingly.
- Adjust imports and rename packages for yield param builders and responses.

This centralizes yield logic, clarifies service boundaries, and prepares for further CAAS-related work.
