+++ src/main/java/exchange/notbank/product/ProductResponseAdapter.java
@@ -11,7 +11,6 @@ import exchange.notbank.core.ErrorHandler;
 import exchange.notbank.core.NotbankException;
 import exchange.notbank.core.responses.DataResponse;
 import exchange.notbank.product.responses.Product;
-import exchange.notbank.product.responses.YieldProduct;
 import exchange.notbank.product.responses.VerificationLevelConfig;
 
 import io.vavr.control.Either;
@@ -20,7 +19,6 @@ public class ProductResponseAdapter {
   private final ErrorHandler errorHandler;
   private final JsonAdapter<Product> productJsonAdapter;
   private final JsonAdapter<List<Product>> productsJsonAdapter;
-  private final JsonAdapter<DataResponse<List<YieldProduct>>> yieldProductsJsonAdapter;
   private final JsonAdapter<VerificationLevelConfig> verificationLevelConfigJsonAdapter;
 
   public ProductResponseAdapter(Moshi moshi) {
@@ -28,11 +26,6 @@ public class ProductResponseAdapter {
     this.productJsonAdapter = moshi.adapter(Product.class);
     ParameterizedType productListType = Types.newParameterizedType(List.class, Product.class);
     this.productsJsonAdapter = moshi.adapter(productListType);
-    ParameterizedType yieldProductListType = Types.newParameterizedType(List.class, YieldProduct.class);
-    ParameterizedType yieldProductListResponseType = Types.newParameterizedType(
-      DataResponse.class, yieldProductListType
-    );
-    this.yieldProductsJsonAdapter = moshi.adapter(yieldProductListResponseType);
     this.verificationLevelConfigJsonAdapter = moshi.adapter(VerificationLevelConfig.class);
   }
 
@@ -50,10 +43,6 @@ public class ProductResponseAdapter {
 
   public Either<NotbankException, List<Product>> toProductList(String jsonStr) {
     return handle(jsonStr, productsJsonAdapter);
-  }
-
-  public Either<NotbankException, List<YieldProduct>> toYieldProductList(String jsonStr) {
-    return handle(jsonStr, yieldProductsJsonAdapter).map(response -> response.data);
   }
 
   public Either<NotbankException, VerificationLevelConfig> toVerificationLevelConfig(String response) {
