+++ src/main/java/exchange/notbank/product/constants/Endpoints.java
@@ -4,5 +4,4 @@ public class Endpoints {
   public static final String GET_PRODUCTS = "GetProducts";
   public static final String GET_PRODUCT = "GetProduct";
   public static final String GET_VERIFICATION_LEVEL_CONFIG = "GetVerificationLevelConfig";
-  public static final String GET_YIELD_PRODUCTS = "yield/products";
 }
