+++ src/main/java/exchange/notbank/wallet/constants/Endpoints.java
@@ -13,6 +13,4 @@ public class Endpoints {
   public static final String FIAT_WITHDRAW = "wallet/fiat/withdrawal";
   public static final String TRANSFER_FUNDS = "wallet/transfer-funds";
   public static final String TRANSACTIONS = "wallet/transactions";
-  public static final String YIELD_WITHDRAW = "yield/withdraw";
-  public static final String YIELD_DEPOSIT = "yield/deposit";
 }
