+++ src/main/java/exchange/notbank/yield/constants/Endpoints.java
@@ -1,0 +1,9 @@
+package exchange.notbank.yield.constants;
+
+public class Endpoints {
+
+  public static final String YIELD_WITHDRAW = "yield/withdraw";
+  public static final String YIELD_DEPOSIT = "yield/deposit";
+  public static final String GET_YIELD_PRODUCTS = "yield/products";
+
+}
