+++ src/main/java/exchange/notbank/yield/responses/YieldProduct.java
@@ -1,4 +1,4 @@
-package exchange.notbank.product.responses;
+package exchange.notbank.yield.responses;
 
 import com.squareup.moshi.Json;
