+++ src/test/java/exchange/notbank/product/ProductServiceTest.java
@@ -8,7 +8,6 @@ import exchange.notbank.NotbankClient;
 import exchange.notbank.TestHelper;
 import exchange.notbank.product.paramBuilders.GetProductParamBuilder;
 import exchange.notbank.product.paramBuilders.GetProductsParamBuilder;
-import exchange.notbank.product.paramBuilders.GetYieldProductsParamBuilder;
 import exchange.notbank.product.paramBuilders.GetVerificationLevelConfigParamBuilder;
 
 public class ProductServiceTest {
@@ -60,12 +59,6 @@ public class ProductServiceTest {
   public void getVerificationLevelConfigParamBuilder() {
     var futureResponse = client.getProductService()
         .getVerificationLevelConfig(new GetVerificationLevelConfigParamBuilder(credentials.accountId));
-    TestHelper.checkNoError(futureResponse);
-  }
-
-  @Test
-  public void getYieldProducts() {
-    var futureResponse = client.getProductService().getYieldProducts(new GetYieldProductsParamBuilder());
     TestHelper.checkNoError(futureResponse);
   }
